Parse `max_htlc_value_in_flight_msat` as an unsigned long (closes #103) (#109)
* handle overflow for field 'max-htlc-value-in-flight'
* moved serialization of uint64 to codecs class
* moved constraint check
* fix configuration (was using hard coded values)
* using UInt64.MaxValue in test constants
* added a default handler for 'ok' messages in Channel

+++ eclair-core/src/test/scala/fr/acinq/eclair/TestConstants.scala
@@ -30,7 +30,7 @@ object TestConstants {
       globalFeatures = "",
       localFeatures = "00",
       dustLimitSatoshis = 500,
-      maxHtlcValueInFlightMsat = 150000000,
+      maxHtlcValueInFlightMsat = UInt64(150000000),
       maxAcceptedHtlcs = 100,
       expiryDeltaBlocks = 144,
       htlcMinimumMsat = 0,
@@ -76,7 +76,7 @@ object TestConstants {
       globalFeatures = "",
       localFeatures = "00", // no announcement
       dustLimitSatoshis = 1000,
-      maxHtlcValueInFlightMsat = Long.MaxValue, // Bob has no limit on the combined max value of in-flight htlcs
+      maxHtlcValueInFlightMsat = UInt64.MaxValue, // Bob has no limit on the combined max value of in-flight htlcs
       maxAcceptedHtlcs = 30,
       expiryDeltaBlocks = 144,
       htlcMinimumMsat = 1000,
